Upgrade on datarecord: not compatible with old database

+++ backend/service/data-record-service.ts
@@ -1,9 +1,23 @@
 import { getDataSource } from "../index";
-import { Repository } from "typeorm";
-import { DataRecord } from '../entity';
+import { EntityManager, Repository } from "typeorm";
+import { DataRecord, DataRecordGroup } from '../entity';
 import { DIProperty, Injectable } from '../lib/di';
 import { UserService } from "./user-service";
 import createError from "http-errors";
+
+
+interface ServiceRepositories {
+    readonly dataRepo:  Repository<DataRecord>;
+    readonly groupRepo: Repository<DataRecordGroup>;
+};
+
+function EntityManager2Repos(em: EntityManager): ServiceRepositories
+{
+    return {
+        dataRepo:  em.getRepository(DataRecord),
+        groupRepo: em.getRepository(DataRecordGroup),
+    };
+}
 
 
 @Injectable({
@@ -11,11 +25,13 @@ import createError from "http-errors";
 })
 export class DataRecordService {
     private drRepo: Repository<DataRecord>;
+    private groupRepo: Repository<DataRecordGroup>;
     @DIProperty(UserService)
     private userService: UserService;
 
     constructor() {
         this.drRepo = getDataSource().getRepository(DataRecord);
+        this.groupRepo = getDataSource().getRepository(DataRecordGroup);
     }
 
     public async insert(username: string, group: string, data: string)
@@ -23,11 +39,69 @@ export class DataRecordService {
         const user = await this.userService.getUser(username);
         if (!user)
             throw createError(404, `User ${username} not found`);
-        const dr = new DataRecord();
-        dr.data = data;
-        dr.dgroup = group;
-        dr.user = Promise.resolve(user);
-        this.drRepo.save(dr);
+
+        await getDataSource().transaction(async (em: EntityManager) => {
+            const repos = EntityManager2Repos(em);
+
+            let dgroup = await repos.groupRepo.findOne({
+                where: {
+                    dgroup: group,
+                    userId: user.id
+                }
+            });
+            if (dgroup == null) {
+                dgroup = new DataRecordGroup();
+                dgroup.dgroup = group;
+                dgroup.userId = user.id;
+            }
+            dgroup = await repos.groupRepo.save(dgroup);
+
+            const dr = new DataRecord();
+            dr.data = data;
+            dr.groupId = dgroup.id;
+            await repos.dataRepo.save(dr);
+        });
+    }
+
+    public async insertListofDatas(username: string, list: {group: string; data: string}[])
+    {
+        const user = await this.userService.getUser(username);
+        if (!user)
+            throw createError(404, `User ${username} not found`);
+
+        await getDataSource().transaction(async (em: EntityManager) => {
+            const repos = EntityManager2Repos(em);
+            const groupcache = {};
+            const saveDatas: DataRecord[] = [];
+
+            for (const gd of list) {
+                let dgroup: DataRecordGroup = groupcache[gd.group];
+                if (dgroup == null) {
+                    dgroup = await repos.groupRepo.findOne({
+                        where: {
+                            dgroup: gd.group,
+                            userId: user.id
+                        }
+                    });
+
+                    if (dgroup == null) {
+                        dgroup = new DataRecordGroup();
+                        dgroup.dgroup = gd.group;
+                        dgroup.userId = user.id;
+                        dgroup.user = Promise.resolve(user);
+                    }
+                    dgroup = await repos.groupRepo.save(dgroup);
+                    groupcache[gd.group] = dgroup;
+                }
+
+                const dr = new DataRecord();
+                dr.data = gd.data;
+                dr.groupId = dgroup.id;
+                saveDatas.push(dr);
+            }
+
+            await repos.dataRepo.save(saveDatas);
+        });
     }
     
     public async getGroups(username: string)
@@ -37,10 +111,9 @@ export class DataRecordService {
             throw createError(404, `User ${username} not found`);
 
         const groups = await 
-            this.drRepo
+            this.groupRepo
             .createQueryBuilder("dr")
             .select("dr.dgroup")
-            .distinct(true)
             .where("dr.user = :user", { user: user.id })
             .getRawMany();
         return groups.map(d => d["dr_dgroup"]);
@@ -53,10 +126,23 @@ export class DataRecordService {
         if (!user)
             throw createError(404, `User ${username} not found`);
 
+        const dgroup = await this.groupRepo.findOne({
+            where: {
+                userId: user.id,
+                dgroup: group,
+            }
+        });
+        if (!dgroup) {
+            return {
+                data: [],
+                count: 0
+            }
+        }
+
         const skip = (pageno ? pageno - 1 : 0) * pagesize;
 		const [result, total] = await this.drRepo.findAndCount(
 			{
-				where: { dgroup: group, userId: user.id }, order: { id: "ASC" },
+                where: { groupId: dgroup.id }, order: { id: "ASC" },
 				take: pagesize,
 				skip: skip
 			}
@@ -74,9 +160,19 @@ export class DataRecordService {
         if (!user)
             throw new createError.NotFound(`User ${username} not found`);
 
+        const dgroup = await this.groupRepo.findOne({
+            where: {
+                userId: user.id,
+                dgroup: group,
+            }
+        });
+        if (!dgroup) {
+            return [];
+        }
+
         const result = await this.drRepo.find(
             {
-                where: { dgroup: group, userId: user.id },
+                where: { groupId: dgroup.id },
                 order: { id: "ASC" },
                 skip: skip
             });
@@ -89,11 +185,11 @@ export class DataRecordService {
         if (!user)
             throw createError(404, `User ${username} not found`);
 
-        await this.drRepo
+        await this.groupRepo
             .createQueryBuilder("dr")
             .delete()
             .where("dgroup = :group", { group: group })
-            .andWhere("user = :user", { user: user.id })
+            .andWhere("userId = :userId", { userId: user.id })
             .execute();
     }
 }
